main.py : moved bulk of code to process_profiles.py per your suggest process_profiles.py : made changes per your comments to recent PR
